+++ src/view/ncl_viewer.cpp
@@ -103,10 +103,10 @@ void NCL_Viewer::_display_all()
 
 void NCL_Viewer::_display_Xplane(int xs)
 {
-    float s1, t1, s2, t2;
-    float x1, y1, z1, y2, z2;
+    double s1, t1, s2, t2;
+    double x1, y1, z1, y2, z2;
 
-    float scl = 1.0;
+    double scl = 1.0;
   //int nxs = xs + geometry->get_nx();
     unsigned int tid;
 
@@ -197,10 +197,10 @@ void NCL_Viewer::_display_Xplane(int xs)
 
 void NCL_Viewer::_display_Yplane(int ys)
 {
-    float s1, t1, s2, t2;
-    float x1, y1, z1, x2, z2;
+    double s1, t1, s2, t2;
+    double x1, y1, z1, x2, z2;
 
-    float scl = 1.0;
+    double scl = 1.0;
 
     unsigned int tid;
 
@@ -287,10 +287,10 @@ void NCL_Viewer::_display_Yplane(int ys)
 
 void NCL_Viewer::_display_Zplane(int zs)
 {
-    float s1, t1, s2, t2;
-    float x1, y1, z1, x2, y2;
+    double s1, t1, s2, t2;
+    double x1, y1, z1, x2, y2;
 
-    float scl = 1.0;
+    double scl = 1.0;
 
     unsigned int tid;
 
@@ -396,16 +396,16 @@ void NCL_Viewer::_display_with_bump()
 void NCL_Viewer::_display_Xplane_with_bump(int xs)
 {
     int   j, k, m;
-    float bs, bf;
-    float delty, deltz;
-    float delty2;
-    float sx, sy, sz;
+    double bs, bf;
+    double delty, deltz;
+    double delty2;
+    double sx, sy, sz;
 
     float minval;
     float *pltvar = NULL;
 
-    float s1, t1, s2, t2;
-    float x1, y1, z1, x2, y2, z2;
+    double s1, t1, s2, t2;
+    double x1, y1, z1, x2, y2, z2;
 
     int _nx = geometry->get_nx();
     int _ny = geometry->get_ny();
@@ -533,16 +533,16 @@ void NCL_Viewer::_display_Xplane_with_bump(int xs)
 void NCL_Viewer::_display_Yplane_with_bump(int ys)
 {
     int   i, k, m;
-    float bs, bf;
-    float deltx, deltz;
-    float deltx2;
-    float sx, sy, sz;
+    double bs, bf;
+    double deltx, deltz;
+    double deltx2;
+    double sx, sy, sz;
 
     float minval;
     float *pltvar = NULL;
 
-    float s1, t1, s2, t2;
-    float x1, y1, z1, x2, y2, z2;
+    double s1, t1, s2, t2;
+    double x1, y1, z1, x2, y2, z2;
 
     int _nx = geometry->get_nx();
     int _ny = geometry->get_ny();
@@ -670,15 +670,15 @@ void NCL_Viewer::_display_Yplane_with_bump(int ys)
 void NCL_Viewer::_display_Zplane_with_bump(int zs)
 {
     int   i, j, n;
-    float bs, bf;
-    float deltx, delty;
-    float sx, sy, sz;
+    double bs, bf;
+    double deltx, delty;
+    double sx, sy, sz;
 
     float minval;
-    double *pltvar = NULL;
+    float *pltvar = NULL;
 
-    float s1, t1, s2, t2;
-    float x1, y1, z1, x2, y2, z2;
+    double s1, t1, s2, t2;
+    double x1, y1, z1, x2, y2, z2;
 
     unsigned int tid;
 
@@ -844,14 +844,14 @@ void NCL_Viewer::_display_Xplane_on_height_surface(int xs)
 {
     int j, k;
   //int m;
-    float *height = NULL;
+    double *height = NULL;
   //float *values = NULL;
 
-    float deltx, delty, deltz, delty2;
-    float sx, sy, sz;
+    double deltx, delty, deltz, delty2;
+    double sx, sy, sz;
 
-    float s1, t1, s2, t2;
-    float x1, y1, z1, y2, z2;
+    double s1, t1, s2, t2;
+    double x1, y1, z1, y2, z2;
 
     int nx = geometry->get_nx();
     int ny = geometry->get_ny();
@@ -861,7 +861,7 @@ void NCL_Viewer::_display_Xplane_on_height_surface(int xs)
 
   //values = geometry->get_lev();
 
-    height = new float[nz * ny];
+    height = new double[nz * ny];
 
     for(k = 0; k < nz; ++k)
     {
@@ -869,7 +869,7 @@ void NCL_Viewer::_display_Xplane_on_height_surface(int xs)
       {
         //m = xs + (j + k * ny) * nx;
         //height[j + k * ny] = values[m];
-          height[j + k * ny] = (float) k / nz - 0.5;
+          height[j + k * ny] = (double) k / nz - 0.5;
       }
     }
 
@@ -981,15 +981,15 @@ void NCL_Viewer::_display_Yplane_on_height_surface(int ys)
 {
     int i, k;
   //int m;
-    float *height = NULL;
+    double *height = NULL;
   //float *values = NULL;
 
-    float deltx,  delty, deltz;
-    float deltx2;
-    float sx, sy, sz;
+    double deltx,  delty, deltz;
+    double deltx2;
+    double sx, sy, sz;
 
-    float s1, t1, s2, t2;
-    float x1, y1, z1, x2, z2;
+    double s1, t1, s2, t2;
+    double x1, y1, z1, x2, z2;
 
     int nx = geometry->get_nx();
     int ny = geometry->get_ny();
@@ -999,7 +999,7 @@ void NCL_Viewer::_display_Yplane_on_height_surface(int ys)
 
   //values = geometry->get_lev();
 
-    height = new float[nz * nx];
+    height = new double[nz * nx];
 
     for(k = 0; k < nz; ++k)
     {
@@ -1007,7 +1007,7 @@ void NCL_Viewer::_display_Yplane_on_height_surface(int ys)
         {
           //m = i + (ys + k * ny) * nx;
           //height[i + k * nx] = values[m];
-            height[i + k * nx] = (float) k / nz - 0.5;
+            height[i + k * nx] = (double) k / nz - 0.5;
         }
     }
 
@@ -1116,14 +1116,14 @@ void NCL_Viewer::_display_Zplane_on_height_surface(int zs)
     int i, j;
   //int n;
   //float *values = NULL;
-    float height = 0.0;
+    double height = 0.0;
 
-    float deltx, delty;
-    float sx, sy;
-  //float sz;
+    double deltx, delty;
+    double sx, sy;
+  //double sz;
 
-    float s1, t1, s2, t2;
-    float x1, y1, z1, x2, y2, z2;
+    double s1, t1, s2, t2;
+    double x1, y1, z1, x2, y2, z2;
 
     int nx = geometry->get_nx();
     int ny = geometry->get_ny();
@@ -1151,7 +1151,7 @@ void NCL_Viewer::_display_Zplane_on_height_surface(int zs)
   //n = zs * nx * ny;
 
   //height = &values[n];
-    height = (float) zs / nz - 0.5;
+    height = (double) zs / nz - 0.5;
 
   //sz = (contour->get_zEndX() - contour->get_zStartX()) / nz;
     z1 = contour->get_zStartX() + height;
@@ -1277,14 +1277,14 @@ void NCL_Viewer::_display_Xplane_on_sphere(int xs)
 {
     int j, k, n;
   //int m;
-    float *height = NULL;
+    double *height = NULL;
   //float *values = NULL;
 
-    float delty, deltz;
-    float sy, sz;
+    double delty, deltz;
+    double sy, sz;
 
-    float s1, t1;
-    float x1, y1, z1;
+    double s1, t1;
+    double x1, y1, z1;
 
     unsigned int tid;
 
@@ -1303,7 +1303,7 @@ void NCL_Viewer::_display_Xplane_on_sphere(int xs)
 
   //values = geometry->get_lev();
 
-    height = new float[ny * nz];
+    height = new double[ny * nz];
 
     for(k = 0; k < nz; ++k)
     {
@@ -1311,7 +1311,7 @@ void NCL_Viewer::_display_Xplane_on_sphere(int xs)
         {
           //m = xs + (j + k * ny) * nx;
           //height[j + k * ny] = values[m];
-            height[j + k * ny] = (float) k;
+            height[j + k * ny] = (double) k;
         }
     }
 
@@ -1359,14 +1359,14 @@ void NCL_Viewer::_display_Yplane_on_sphere(int ys)
 {
     int i, k, n;
   //int i, k, n;
-    float *height = NULL;
+    double *height = NULL;
   //float *values = NULL;
 
-    float deltx, deltz;
-    float sx, sz;
+    double deltx, deltz;
+    double sx, sz;
 
-    float s1, t1;
-    float x1, y1, z2;
+    double s1, t1;
+    double x1, y1, z2;
 
     unsigned int tid;
 
@@ -1384,7 +1384,7 @@ void NCL_Viewer::_display_Yplane_on_sphere(int ys)
     lat = geometry->get_lat();
 
   //values = geometry->get_lev();
-    height = new float[nx * nz];
+    height = new double[nx * nz];
 
     for(k = 0; k < nz; ++k)
     {
@@ -1392,7 +1392,7 @@ void NCL_Viewer::_display_Yplane_on_sphere(int ys)
         {
           //m = i + (ys + k * ny) * nx;
           //height[i + k * nx] = values[m];
-            height[i + k * nx] = (float) k;
+            height[i + k * nx] = (double) k;
         }
     }
 
@@ -1440,14 +1440,13 @@ void NCL_Viewer::_display_Zplane_on_sphere(int zs)
 {
     int i, j, n;
     double height;
-  //double *height = NULL;
-  //double *values = NULL;
+  //float *values = NULL;
 
-    float deltx, delty;
-    float sx, sy, sz;
+    double deltx, delty;
+    double sx, sy, sz;
 
-    float s1, t1;
-    float x1, y1, z1;
+    double s1, t1;
+    double x1, y1, z1;
 
     unsigned int tid;
 
@@ -1522,10 +1521,10 @@ void NCL_Viewer::_display_Global_on_sphere()
 {
     int i, j;
 
-    float deltx, delty;
+    double deltx, delty;
 
-    float s1, t1, t2;
-    float x1, y1, y2;
+    double s1, t1, t2;
+    double x1, y1, y2;
 
     unsigned int tid;
 
@@ -1541,7 +1540,7 @@ void NCL_Viewer::_display_Global_on_sphere()
 
     for(j = -90; j < 90; j += 5)
     {
-        y1 = (float) j;
+        y1 = (double) j;
         y2 = y1 + 5.0;
         t1 = contour->get_yStart() + delty * (0.5 + y1/180.0);
         t2 = contour->get_yStart() + delty * (0.5 + y2/180.0);
@@ -1549,7 +1548,7 @@ void NCL_Viewer::_display_Global_on_sphere()
         glBegin(GL_QUAD_STRIP);
         for(i = 0; i <= 360; i += 5)
         {
-            x1 = (float) i;
+            x1 = (double) i;
             s1 = contour->get_xStart() + deltx*x1/360.0;
             _Vertex(x1, y1, 0.505, 1.0 - s1, 1.0 - t1);
             _Vertex(x1, y2, 0.505, 1.0 - s1, 1.0 - t2);
@@ -1572,7 +1571,7 @@ void NCL_Viewer::set_geometry(Geometry *gm)
     contour->set_nz(geometry->get_nz());
 }
 
-void NCL_Viewer::setup(string vn, double *var)
+void NCL_Viewer::setup(string vn, float *var)
 {
     _varname  = vn;
     _var = var;
@@ -1599,24 +1598,24 @@ void NCL_Viewer::set_OP(bool s)
 }
 
 //Draw vertex in polar coordinates
-void NCL_Viewer::_Vertex(float lon, float lat, float r, float s, float t)
+void NCL_Viewer::_Vertex(double lon, double lat, double r, double s, double t)
 {
-    float x = r*Cos(lon)*Cos(lat);
-    float y =          r*Sin(lat);
-    float z = r*Sin(lon)*Cos(lat);
+    double x = r*Cos(lon)*Cos(lat);
+    double y =          r*Sin(lat);
+    double z = r*Sin(lon)*Cos(lat);
   //glNormal3d(x,y,z);
     glTexCoord2d(s, t);
     glVertex3d(x,y,z);
 }
 
-void NCL_Viewer::_point(float x, float y, float z, float s, float t)
+void NCL_Viewer::_point(double x, double y, double z, double s, double t)
 {
     glTexCoord2d(s, 1.0 - t);
     glVertex3d(x,y,z);
 }
 
-void NCL_Viewer::_xQuad(float x1, float y1, float z1, float y2, float z2,
-                    float s1, float t1, float s2, float t2)
+void NCL_Viewer::_xQuad(double x1, double y1, double z1, double y2, double z2,
+                        double s1, double t1, double s2, double t2)
 {
     _point(x1, y1, z1, s1, t1);
     _point(x1, y2, z1, s2, t1);
@@ -1624,8 +1623,8 @@ void NCL_Viewer::_xQuad(float x1, float y1, float z1, float y2, float z2,
     _point(x1, y1, z2, s1, t2);
 }
 
-void NCL_Viewer::_yQuad(float x1, float y1, float z1, float x2, float z2,
-                    float s1, float t1, float s2, float t2)
+void NCL_Viewer::_yQuad(double x1, double y1, double z1, double x2, double z2,
+                        double s1, double t1, double s2, double t2)
 {
     _point(x1, y1, z1, s1, t1);
     _point(x2, y1, z1, s2, t1);
@@ -1633,8 +1632,8 @@ void NCL_Viewer::_yQuad(float x1, float y1, float z1, float x2, float z2,
     _point(x1, y1, z2, s1, t2);
 }
 
-void NCL_Viewer::_zQuad(float x1, float y1, float x2, float y2, float z2,
-                    float s1, float t1, float s2, float t2)
+void NCL_Viewer::_zQuad(double x1, double y1, double x2, double y2, double z2,
+                        double s1, double t1, double s2, double t2)
 {
     _point(x1, y1, z2, s1, t1);
     _point(x2, y1, z2, s2, t1);
